Rename mapping package to coloredmodel
