+++ web-ui/src/main/java/org/archcnl/domain/input/visualization/ConceptVisualizer.java
@@ -11,6 +11,10 @@ import org.archcnl.domain.common.RelationManager;
 import org.archcnl.domain.common.conceptsandrelations.CustomConcept;
 import org.archcnl.domain.common.conceptsandrelations.andtriplets.triplet.Variable;
 import org.archcnl.domain.input.model.mappings.ConceptMapping;
+import org.archcnl.domain.input.visualization.coloredmodel.ColorState;
+import org.archcnl.domain.input.visualization.coloredmodel.ColoredTriplet;
+import org.archcnl.domain.input.visualization.coloredmodel.ColoredVariant;
+import org.archcnl.domain.input.visualization.coloredmodel.MappingPreprocessor;
 import org.archcnl.domain.input.visualization.diagram.ConceptBlock;
 import org.archcnl.domain.input.visualization.diagram.PlantUmlBlock;
 import org.archcnl.domain.input.visualization.diagram.PlantUmlPart;
@@ -19,10 +23,6 @@ import org.archcnl.domain.input.visualization.diagram.elements.PlantUmlElement;
 import org.archcnl.domain.input.visualization.exceptions.MappingToUmlTranslationFailedException;
 import org.archcnl.domain.input.visualization.exceptions.PropertyNotFoundException;
 import org.archcnl.domain.input.visualization.helpers.NamePicker;
-import org.archcnl.domain.input.visualization.mapping.ColorState;
-import org.archcnl.domain.input.visualization.mapping.ColoredTriplet;
-import org.archcnl.domain.input.visualization.mapping.ColoredVariant;
-import org.archcnl.domain.input.visualization.mapping.MappingPreprocessor;
 
 public class ConceptVisualizer extends MappingVisualizer implements PlantUmlBlock {
