+++ web-ui/src/main/java/org/archcnl/domain/input/visualization/MappingTranslator.java
@@ -22,13 +22,13 @@ import org.archcnl.domain.common.conceptsandrelations.andtriplets.triplet.Triple
 import org.archcnl.domain.common.conceptsandrelations.andtriplets.triplet.Variable;
 import org.archcnl.domain.input.model.mappings.ConceptMapping;
 import org.archcnl.domain.input.visualization.diagram.PlantUmlBlock;
-import org.archcnl.domain.input.visualization.diagram.PlantUmlMapper;
+import org.archcnl.domain.input.visualization.coloredmodel.ColorState;
+import org.archcnl.domain.input.visualization.coloredmodel.ColoredTriplet;
+import org.archcnl.domain.input.visualization.diagram.ConceptMapper;
 import org.archcnl.domain.input.visualization.diagram.PlantUmlPart;
 import org.archcnl.domain.input.visualization.diagram.connections.PlantUmlConnection;
 import org.archcnl.domain.input.visualization.diagram.elements.PlantUmlElement;
 import org.archcnl.domain.input.visualization.exceptions.MappingToUmlTranslationFailedException;
-import org.archcnl.domain.input.visualization.mapping.ColorState;
-import org.archcnl.domain.input.visualization.mapping.ColoredTriplet;
 
 public class MappingTranslator {
 
@@ -122,7 +122,7 @@ public class MappingTranslator {
                                 colorState);
                 elementMap.put(variable, visualizer);
             } else {
-                PlantUmlElement element = PlantUmlMapper.createElement(elementType, variable);
+                PlantUmlElement element = ConceptMapper.createElement(elementType, variable);
                 element.setColorState(colorState);
                 elementMap.put(variable, element);
             }
