+++ web-ui/src/main/java/org/archcnl/domain/input/visualization/diagram/elements/Field.java
@@ -5,9 +5,9 @@ import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
 import org.archcnl.domain.common.conceptsandrelations.andtriplets.triplet.Variable;
+import org.archcnl.domain.input.visualization.coloredmodel.ColorState;
 import org.archcnl.domain.input.visualization.diagram.elements.containers.ModifierContainer;
 import org.archcnl.domain.input.visualization.exceptions.PropertyNotFoundException;
-import org.archcnl.domain.input.visualization.mapping.ColorState;
 
 public class Field extends PlantUmlElement {
