+++ web-ui/src/main/java/org/archcnl/domain/input/visualization/diagram/elements/PlantUmlElement.java
@@ -5,10 +5,10 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
 import org.archcnl.domain.common.conceptsandrelations.andtriplets.triplet.Variable;
+import org.archcnl.domain.input.visualization.coloredmodel.ColorState;
 import org.archcnl.domain.input.visualization.diagram.PlantUmlBlock;
 import org.archcnl.domain.input.visualization.exceptions.PropertyNotFoundException;
 import org.archcnl.domain.input.visualization.helpers.NamePicker;
-import org.archcnl.domain.input.visualization.mapping.ColorState;
 
 public abstract class PlantUmlElement implements PlantUmlBlock {
