+++ web-ui/src/main/java/org/archcnl/domain/input/visualization/diagram/elements/PrimitiveType.java
@@ -2,8 +2,8 @@ package org.archcnl.domain.input.visualization.diagram.elements;
 
 import java.util.Optional;
 import org.archcnl.domain.common.conceptsandrelations.andtriplets.triplet.Variable;
+import org.archcnl.domain.input.visualization.coloredmodel.ColorState;
 import org.archcnl.domain.input.visualization.exceptions.PropertyNotFoundException;
-import org.archcnl.domain.input.visualization.mapping.ColorState;
 
 public class PrimitiveType extends PlantUmlElement {
