+++ web-ui/src/main/java/org/archcnl/domain/input/visualization/rules/CardinalityRuleVisualizer.java
@@ -7,8 +7,8 @@ import java.util.regex.Pattern;
 import org.archcnl.domain.common.ConceptManager;
 import org.archcnl.domain.common.RelationManager;
 import org.archcnl.domain.input.model.architecturerules.ArchitectureRule;
+import org.archcnl.domain.input.visualization.coloredmodel.ColorState;
 import org.archcnl.domain.input.visualization.exceptions.MappingToUmlTranslationFailedException;
-import org.archcnl.domain.input.visualization.mapping.ColorState;
 
 public class CardinalityRuleVisualizer extends RuleVisualizer {
