+++ web-ui/src/main/java/org/archcnl/domain/input/visualization/rules/ConditionalRuleVisualizer.java
@@ -8,9 +8,8 @@ import org.archcnl.domain.common.ConceptManager;
 import org.archcnl.domain.common.RelationManager;
 import org.archcnl.domain.common.conceptsandrelations.Relation;
 import org.archcnl.domain.input.model.architecturerules.ArchitectureRule;
+import org.archcnl.domain.input.visualization.coloredmodel.ColorState;
 import org.archcnl.domain.input.visualization.exceptions.MappingToUmlTranslationFailedException;
-import org.archcnl.domain.input.visualization.helpers.RuleHelper;
-import org.archcnl.domain.input.visualization.mapping.ColorState;
 
 public class ConditionalRuleVisualizer extends RuleVisualizer {
