+++ web-ui/src/main/java/org/archcnl/domain/input/visualization/rules/RulePredicate.java
@@ -3,8 +3,8 @@ package org.archcnl.domain.input.visualization.rules;
 import java.util.Optional;
 import org.archcnl.domain.common.conceptsandrelations.Relation;
 import org.archcnl.domain.common.conceptsandrelations.andtriplets.triplet.Variable;
-import org.archcnl.domain.input.visualization.mapping.ColorState;
-import org.archcnl.domain.input.visualization.mapping.ColoredTriplet;
+import org.archcnl.domain.input.visualization.coloredmodel.ColorState;
+import org.archcnl.domain.input.visualization.coloredmodel.ColoredTriplet;
 
 public class RulePredicate {
