+++ web-ui/src/main/java/org/archcnl/domain/input/visualization/rules/RuleVisualizer.java
@@ -20,15 +20,14 @@ import org.archcnl.domain.common.conceptsandrelations.andtriplets.triplet.Variab
 import org.archcnl.domain.input.model.architecturerules.ArchitectureRule;
 import org.archcnl.domain.input.visualization.MappingTranslator;
 import org.archcnl.domain.input.visualization.Visualizer;
+import org.archcnl.domain.input.visualization.coloredmodel.ColoredTriplet;
+import org.archcnl.domain.input.visualization.coloredmodel.ColoredVariant;
+import org.archcnl.domain.input.visualization.coloredmodel.MappingSetter;
 import org.archcnl.domain.input.visualization.diagram.PlantUmlBlock;
 import org.archcnl.domain.input.visualization.diagram.PlantUmlPart;
 import org.archcnl.domain.input.visualization.exceptions.MappingToUmlTranslationFailedException;
 import org.archcnl.domain.input.visualization.helpers.MappingFlattener;
 import org.archcnl.domain.input.visualization.helpers.NamePicker;
-import org.archcnl.domain.input.visualization.helpers.RuleHelper;
-import org.archcnl.domain.input.visualization.mapping.ColoredTriplet;
-import org.archcnl.domain.input.visualization.mapping.ColoredVariant;
-import org.archcnl.domain.input.visualization.mapping.MappingSetter;
 import org.archcnl.domain.input.visualization.rules.VerbPhraseContainer.Connector;
 
 public abstract class RuleVisualizer implements Visualizer {
